requirements.txt
fastapi
pydantic
uvicorn[standard]
psycopg2
flask-sqlalchemy
